Fix: According to MDN at https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input, min and max can take a date-time value.
When that happens, these two validators ends up trying to convert these values into a number, which results in NaN, which causes the validation to fail.

To fix this issue, if the min or max value isNaN and is a date, then simply return null. The validtion of these fields are handled by minDate and maxDate